+++ src/python/DAS/services/runregistry/runregistry_service.py
@@ -12,15 +12,15 @@ import time
 import xmlrpclib
 import DAS.utils.jsonwrapper as json
 from   DAS.services.abstract_service import DASAbstractService
-from   DAS.utils.utils import map_validator, adjust_value
+from   DAS.utils.utils import map_validator, adjust_value, next_day
 
 def convert_datetime(sec):
     """
     Convert seconds since epoch or YYYYMMDD to date format used in RunRegistry
     """
     value = str(sec)
-    if  value == 8: # we got YYYYMMDD
-        return "%s-%s-%s" % (value[:4], value[5:6], value[7:8])
+    if  len(value) == 8: # we got YYYYMMDD
+        return "%s-%s-%s" % (value[:4], value[4:6], value[6:8])
     return time.strftime("%Y-%m-%d", time.gmtime(sec))
     
 def worker(url, query):
@@ -110,12 +110,27 @@ class RunRegistryService(DASAbstractService):
                     _query += "{runNumber} >= %s and {runNumber} <= %s" \
                             % (minrun, maxrun)
             elif key == 'date':
-                if  val.has_key('$in'):
-                    value = val['$in']
+                if  isinstance(val, dict):
+                    if  val.has_key('$in'):
+                        value = val['$in']
+                    elif val.has_key('$between'):
+                        value = val['$between']
+                    else:
+                        msg = 'Unable to get the value from %s=%s' \
+                                % (key, val) 
+                        raise Exception(msg)
+                else:
+                     value = [val, next_day(val)]
+                try:
                     date1 = convert_datetime(value[0])
                     date2 = convert_datetime(value[-1])
-                    run_time = '>= %s and < %s' % (date1, date2)
-                    _query = {'runStartTime': run_time}
+                except:
+                    msg = 'Unable to convert to datetime format, %s' \
+                        % value
+                    raise Exception(msg)
+                run_time = '>= %s and < %s' % (date1, date2)
+                _query = {'runStartTime': run_time}
+        print "\n### query", _query
         if  not _query:
             msg = 'Unable to match input parameters with input query'
             raise Exception(msg)