+++ src/python/DAS/utils/utils.py
@@ -91,6 +91,27 @@ def get_http_expires(data):
         time.strptime(expire, '%a, %d %b %Y %H:%M:%S %Z') )
     return expire
 
+def to_seconds(das_date):
+    """Convert provided DAS date into seconds since epoch"""
+    tstamp = str(das_date)
+    if  len(tstamp) != 8:
+        raise Exception('Provided date does not confirm DAS date format')
+    year  = tstamp[:4]
+    month = tstamp[4:6]
+    day   = tstamp[6:8]
+    obj   = datetime.date(int(year), int(month), int(day))
+    return time.mktime(obj.timetuple())
+
+def next_day(das_date):
+    """Return next date provided DAS date""" 
+    next  = to_seconds(das_date) + 24*60*60
+    return int(time.strftime("%Y%m%d", time.gmtime(next)))
+    
+def prev_day(das_date):
+    """Return previous date provided DAS date""" 
+    next  = to_seconds(das_date) - 24*60*60
+    return int(time.strftime("%Y%m%d", time.gmtime(next)))
+
 def expire_timestamp(expire):
     """
     Return expire timestamp. The input parameter expire can be in a form of